Move session-scoped fixturtes to conftest (globally available)
